backend/requirements.txt
fastapi
uvicorn
google-cloud-aiplatform
google-cloud-bigquery
pydantic
python-dotenv
pydantic-settings
